Compiles and tests start to run but panic
