+++ cedar-drt/fuzz/fuzz_targets/abac-type-directed.rs
@@ -143,8 +143,8 @@ fuzz_target!(|input: FuzzTargetInput| {
     initialize_log();
     let mut policyset = ast::PolicySet::new();
     policyset.add_static(input.policy.into()).unwrap();
-    let java_def_engine =
-        JavaDefinitionalEngine::new().expect("failed to create definitional engine");
+    let lean_def_engine =
+        LeanDefinitionalEngine::new().expect("failed to create definitional engine");
     debug!("Schema: {}\n", input.schema.schemafile_string());
     debug!("Policies: {policyset}\n");
     debug!("Entities: {}\n", input.entities);
@@ -157,14 +157,14 @@ fuzz_target!(|input: FuzzTargetInput| {
     for request in input.requests.into_iter().map(Into::into) {
         debug!("Request : {request}");
         let (rust_res, total_dur) = time_function(|| {
-            run_auth_test(&java_def_engine, &request, &policyset, &original_entities)
+            run_auth_test(&lean_def_engine, &request, &policyset, &original_entities)
         });
 
         if let Some(ref entities) = cached_entities {
             match entities {
                 Ok(entities) => {
                     let (cached_rust_res, _total_dur) = time_function(|| {
-                        run_auth_test(&java_def_engine, &request, &policyset, entities)
+                        run_auth_test(&lean_def_engine, &request, &policyset, entities)
                     });
                     assert_eq!(rust_res, cached_rust_res);
                 }
