+++ cedar-drt/src/lib.rs
@@ -16,11 +16,11 @@
 
 // #![forbid(unsafe_code)]
 mod cedar_test_impl;
-mod dafny_java_impl;
+mod definitional_request_types;
 mod lean_impl;
 mod logger;
 
+pub use definitional_request_types::*;
 pub use cedar_test_impl::*;
-pub use dafny_java_impl::*;
 pub use lean_impl::*;
 pub use logger::*;
