+++ cedar-drt/tests/integration_tests.rs
@@ -69,15 +69,8 @@ fn run_integration_tests(custom_impl: &dyn CustomCedarImpl) {
 }
 
 #[test]
-fn integration_tests_on_java_def_impl() {
-    let java_def_impl =
-        JavaDefinitionalEngine::new().expect("failed to create definitional engine");
-    run_integration_tests(&java_def_impl)
-}
-
-#[test]
 fn integration_tests_on_lean_def_impl() {
     let lean_def_impl =
-        LeanDefinitionalEngine::new();
+        LeanDefinitionalEngine::new().unwrap();
     run_integration_tests(&lean_def_impl)
 }
